+++ package.json
@@ -6,7 +6,7 @@
   "dependencies": {
     "@graasp/chatbox": "github:graasp/graasp-chatbox.git",
     "@graasp/query-client": "github:graasp/graasp-query-client.git",
-    "@graasp/ui": "github:graasp/graasp-ui.git#92/ccLicense",
+    "@graasp/ui": "github:graasp/graasp-ui.git",
     "@graasp/utils": "github:graasp/graasp-utils.git",
     "@material-ui/core": "4.12.3",
     "@material-ui/icons": "5.0.0-beta.4",
